+++ Tests/Integration/MSFT_xDnsClientGlobalSetting.Integration.Tests.ps1
@@ -24,11 +24,34 @@ $CurrentDnsClientGlobalSetting = Get-DnsClientGlobalSetting
 # Using try/finally to always cleanup even if something awful happens.
 try
 {
+    # Import the Common Networking functions
+    Import-Module -Name (Join-Path -Path $script:moduleRoot -ChildPath 'Modules\NetworkingDsc.Common\NetworkingDsc.Common.psm1') -Force
+
+    # Load the ParameterList from the data file.
+    $resourceDataPath = Join-Path `
+        -Path $script:moduleRoot `
+        -ChildPath (Join-Path -Path 'DSCResources' -ChildPath $script:DSCResourceName)
+    $resourceData = Import-LocalizedData `
+        -BaseDirectory $resourceDataPath `
+        -FileName "$($script:DSCResourceName).data.psd1"
+    $parameterList = $resourceData.ParameterList
+
     # Set the DNS Client Global settings to known values
     Set-DnsClientGlobalSetting `
         -SuffixSearchList 'fabrikam.com' `
         -UseDevolution $False `
         -DevolutionLevel 4
+
+    $configData = @{
+        AllNodes = @(
+            @{
+                NodeName              = 'localhost'
+                SuffixSearchList             = 'contoso.com'
+                UseDevolution                = $True
+                DevolutionLevel              = 2
+            }
+        )
+    }
 
     #region Integration Tests
     $ConfigFile = Join-Path -Path $PSScriptRoot -ChildPath "$($script:DSCResourceName).config.ps1"
@@ -38,8 +61,11 @@ try
         #region DEFAULT TESTS
         It 'Should compile without throwing' {
             {
-                & "$($script:DSCResourceName)_Config" -OutputPath $TestDrive
-                Start-DscConfiguration -Path $TestDrive -ComputerName localhost -Wait -Verbose -Force
+                & "$($script:DSCResourceName)_Config" `
+                    -OutputPath $TestDrive `
+                    -ConfigurationData $configData
+                Start-DscConfiguration `
+                    -Path $TestDrive -ComputerName localhost -Wait -Verbose -Force
             } | Should not throw
         }
 
@@ -48,12 +74,20 @@ try
         }
         #endregion
 
-        It 'Should have set the resource and all the parameters should match' {
-            # Get the Rule details
-            $DnsClientGlobalSettingNew = Get-DnsClientGlobalSetting
-            $DnsClientGlobalSettingNew.SuffixSearchList | Should Be $DnsClientGlobalSetting.SuffixSearchList
-            $DnsClientGlobalSettingNew.UseDevolution    | Should Be $DnsClientGlobalSetting.UseDevolution
-            $DnsClientGlobalSettingNew.DevolutionLevel  | Should Be $DnsClientGlobalSetting.DevolutionLevel
+        # Get the DNS Client Global Settings details
+        $DnsClientGlobalSettingNew = Get-DnsClientGlobalSetting
+
+        # Use the Parameters List to perform these tests
+        foreach ($parameter in $parameterList)
+        {
+            $parameterName = $parameter.Name
+            if ($parameterName -ne 'Name') {
+                $parameterSource = (Invoke-Expression -Command "`$($($parameter.source))")
+                $parameterNew = (Invoke-Expression -Command "`$configData.AllNodes[0].$($parameter.name)")
+                It "Should have set the '$parameterName' to '$parameterNew'" {
+                    $parameterSource | Should Be $parameterNew
+                }
+            }
         }
     }
     #endregion